башни

+++ core/src/com/mygdx/game/objects/GameObject.java
@@ -18,7 +18,7 @@ public abstract class GameObject {
 
     Texture texture;
 
-    protected GameObject(String texturePath, int x, int y, int width, int height, short cBits, World world) {
+    protected GameObject(String texturePath, float x, float y, int width, int height, short cBits, World world) {
         this.width = width;
         this.height = height;
         this.cBits = cBits;
